Windows compilation
maybe ;)

+++ src/cpp_odbc/cpp_odbc/multi_value_buffer.h
@@ -3,7 +3,7 @@
  *  @file multi_value_buffer.h
  *  @date 11.04.2014
  *  @author mkoenig
- *  @brief 
+ *  @brief
  *
  *  $LastChangedDate: 2014-11-28 11:59:59 +0100 (Fr, 28 Nov 2014) $
  *  $LastChangedBy: mkoenig $
@@ -17,12 +17,12 @@ namespace cpp_odbc {
 
 struct writable_buffer_element {
 	char * data_pointer;
-	long & indicator;
+	long long & indicator;
 };
 
 struct buffer_element {
 	char const * data_pointer;
-	long const & indicator;
+	long long const & indicator;
 };
 
 /**
@@ -72,7 +72,7 @@ public:
 	 *        StrLen_or_IndPtrs appear in ODBC calls
 	 * @return A pointer to the length/indicator array
 	 */
-	long * indicator_pointer();
+	long long * indicator_pointer();
 
 	/**
 	 * @brief Provides read-only raw access to a contiguous strip of memory which
@@ -80,7 +80,7 @@ public:
 	 *        is NULL).
 	 * @return A pointer to the length/indicator array
 	 */
-	long const * indicator_pointer() const;
+	long long const * indicator_pointer() const;
 
 	/**
 	 * @brief Read/write access to the element with index element_index
@@ -99,7 +99,7 @@ public:
 private:
 	std::size_t element_size_;
 	std::vector<char> data_;
-	std::vector<long> indicators_;
+	std::vector<long long> indicators_;
 };
